Controllers refactored pre-original delete
